# Conflicts:
#	mallplus-admin/src/main/java/com/zscat/mallplus/ApiContext.java
#	mallplus-admin/src/main/java/com/zscat/mallplus/config/MybatisPlusConfig.java

+++ mallplus-admin/src/main/java/com/zscat/mallplus/config/MybatisPlusConfig.java
@@ -52,9 +52,9 @@ public class MybatisPlusConfig {
             @Override
             public Expression getTenantId() {
                 // 从当前系统上下文中取出当前请求的服务商ID，通过解析器注入到SQL中。
-                Integer currentProviderId = apiContext.getCurrentProviderId();
+                Long currentProviderId = apiContext.getCurrentProviderId();
                 if (null == currentProviderId) {
-                    currentProviderId = 5;
+                    currentProviderId = 5l;
                     System.out.println("#1129 getCurrentProviderId error.");
                     throw new RuntimeException("#1129 getCurrentProviderId error.");
                 }
